server/requirements.txt
django
djangorestframework
postgres
channels
pyjwt
